+++ src/routers/AppRouter.jsx
@@ -7,7 +7,7 @@ import { DashBoard } from './DashBoard';
 import AngelInfo from '../containers/angel/Angel';
 import Notifications from '../containers/Notifications/Notifications';
 import RestartPassword from '../containers/Restart/Restart';
-import ChangeInfo from '../containers/ChangeInfo/ChangeInfo';
+import ProfileAndChangeInfo from '../containers/ProfileAndChangeInfo/ProfileAndChangeInfo';
 
 export const AppRouter = () => {
   return (
@@ -15,7 +15,7 @@ export const AppRouter = () => {
       <Switch>
         <Route exact path="/alerta" component={Notifications} />
         <Route exact path="/login" component={Login} />
-        <Route exact path="/profile" component={ChangeInfo} />
+        <Route exact path="/profile" component={ProfileAndChangeInfo} />
         <Route exact path="/registro" component={Register} />
         <Route path="/angel" component={AngelInfo} />
         <Route path="/restart" component={RestartPassword} />
